Re-organize tests
- Now each hooks test has their own file
