+++ plugins/ui/test/deephaven/ui/hooks/test_state.py
@@ -1,0 +1,45 @@
+from operator import itemgetter
+from ..BaseTest import BaseTestCase
+from .render_utils import render_hook
+
+
+class UseStateTestCase(BaseTestCase):
+    def test_state(self):
+        from deephaven.ui.hooks import use_state
+
+        def _test_state(value1: int = 1, value2: int = 2):
+            value1, set_value1 = use_state(value1)
+            value2, set_value2 = use_state(value2)
+            return value1, set_value1, value2, set_value2
+
+        # Initial render
+        render_result = render_hook(_test_state)
+
+        result, rerender = itemgetter("result", "rerender")(render_result)
+        val1, set_val1, val2, set_val2 = result
+
+        self.assertEqual(val1, 1)
+        self.assertEqual(val2, 2)
+
+        # Rerender with new values, but should retain existing state
+        rerender(value1=3, value2=4)
+        result = itemgetter("result")(render_result)
+        val1, set_val1, val2, set_val2 = result
+        self.assertEqual(val1, 1)
+        self.assertEqual(val2, 2)
+
+        # Set to a new value
+        set_val1(3)
+        rerender()
+        result = itemgetter("result")(render_result)
+        val1, set_val1, val2, set_val2 = result
+        self.assertEqual(val1, 3)
+        self.assertEqual(val2, 2)
+
+        # Set other state to a new value
+        set_val2(4)
+        rerender()
+        result = itemgetter("result")(render_result)
+        val1, set_val1, val2, set_val2 = result
+        self.assertEqual(val1, 3)
+        self.assertEqual(val2, 4)
